[tables] Fix crash with position:absolute display:table marquee.
During NGBlockNode::ComputeMinMaxSizes we occasionally set the min_size
to zero, (for example if its a <marquee> element).

This is particularly bad for table layout as we assume that we always
have our min_size being whatever the ComputeMinMaxSizes returned.

This patch breaks the tie between these two constraints, favouring
the table min-constraint.

Bug: 1279129
Change-Id: I8406546ba7b1c514dd194966519735dc9dc4d952